Extract the pipeline definitions out into separate functions
