+++ src/poprox_recommender/recommenders/__init__.py
@@ -1,0 +1,14 @@
+from poprox_recommender.recommenders.nrms import nrms_pipeline
+from poprox_recommender.recommenders.nrms_calibration import locality_calibration_pipeline, topic_calibration_pipeline
+from poprox_recommender.recommenders.nrms_mmr import nrms_mmr_pipeline
+from poprox_recommender.recommenders.nrms_pfar import nrms_pfar_pipeline
+from poprox_recommender.recommenders.nrms_softmax import nrms_softmax_pipeline
+
+__all__ = [
+    "nrms_pipeline",
+    "locality_calibration_pipeline",
+    "topic_calibration_pipeline",
+    "nrms_mmr_pipeline",
+    "nrms_pfar_pipeline",
+    "nrms_softmax_pipeline",
+]
