+++ tests/integration/test_smoke.py
@@ -10,7 +10,7 @@ from poprox_concepts import ArticleSet, Click
 from poprox_concepts.api.recommendations import RecommendationRequest
 from poprox_recommender.config import allow_data_test_failures
 from poprox_recommender.paths import project_root
-from poprox_recommender.recommenders import PipelineLoadError, select_articles
+from poprox_recommender.recommend import PipelineLoadError, select_articles
 
 logger = logging.getLogger(__name__)
